Save Point 1
The game made in the previous semester using Java. This program was inspired by the Chain Reaction game on Mobile platform.  The first iterations only had a "bot" which used randomized numbers to select a grid, with a base function checking by parsing through all grids to make sure there were still grids to be locked. This approach was time consuming, resource intensive and overall bad experience for the user.