Save Point 2
This program was the next iteration of the previous game, with the bare minimum GUI and focus on an efficient and strategic algorithm using Python Language. The algorithm is simple, given the users input, a value is generated such that it is close vicinity while trying to lock maximum grids as possible or to lock the users grids. This approach repeats a few times and then the algorithm systematically parses from the top to find the unlocked grids and lock them.